CDK-964: Allow replacing provided partitioners.
This updates the update validation so that provided partitioners can be
replaced with more specific partitioners. For example, if a version
partition is initially provided but later added to the record data, the
provided partitioner can be replaced with an identity partitioner. This
is the only allowed migration added by this commit.