CDK-938: Add kite-data-s3
This adds both s3n and s3a dataset URIs and a test for each.  Both URI
patterns assume the bucket is the dataset repository.

Until a mock S3 service can be added, this expects real S3 credentials
and a bucket name passed as S3_ID, S3_KEY, and S3_BUCKET environment
variables. If those are not present, then the test will not run. To
disable tests for the hadoop1 and cdh4 profiles, the credentials are
explicitly set to empty.

Validated that the CLI works with S3 URIs.

Conflicts:
	pom.xml
Resolution:
    Conflict with CDH pom changes.
    Need to follow up with a CLOUDERA-BUILD commit to set versions.