CDK-976: Fix test failures caused by default conf changes.
The default conf needs to be set in MR tasks, but always setting the
configuration in InputFormat and OutputFormat messages isn't correct
because the methods aren't necessarily called in a new process. Using
those methods within a test process, for example, broke several CLI
tests because the DefaultConfiguration is shared.

This fixes the problem by adding an init method to DefaultConfiguration
that will set the default conf if it has not already been set. The set
method will always set the conf and prevents init from setting it later.