KITE-1083: Add single jars to DistCache for Hive.
This changes the way jobs submitted by the CLI are configured.
Previously, the entire lib directory for Hive was added to the
distributed cache. This caused long job and task startup times and
exposed some conflicting jar problems. This commit updates the setup so
that individual jars are added for classes needed for interacting with
the Hive MetaStore. In cases where the job is local or the job isn't
interacting with Hive, this doesn't add Hive dependencies to the
distributed cache at all.